requirements.txt
-i https://pypi.org/simple
--extra-index-url https://pypi.org/simple
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.3.5
beautifulsoup4==4.6.1
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
django-cors-headers==2.4.0
django-rest-framework==0.1.0
django-silk==3.0.1
django-webpack-loader==0.6.0
django-webtest==1.9.3
django==2.1
djangorestframework==3.8.2
gprof2dot==2016.10.13
gunicorn==19.9.0
html5lib==1.0.1
idna==2.7
jinja2==2.10
lxml==4.2.4
markupsafe==1.0
more-itertools==4.3.0
pluggy==0.7.1
psycopg2-binary==2.7.5
py==1.6.0
pycodestyle==2.4.0
pygments==2.2.0
pytest-django==3.4.2
pytest==3.8.0
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
requests==2.19.1
six==1.11.0
sqlparse==0.2.4
urllib3==1.23
waitress==1.1.0
webencodings==0.5.1
webob==1.8.2
webtest==2.0.30
